# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Planificacion Horaria",
        "description": "Gestion de horarios de clases",
        "contact": {
            "email": "contacto@tudominio.com"
        },
        "license": {
            "name": "Licencia",
            "url": "https://www.terciarioUrquiza.com.ar/planificacionHoraria"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/aulas": {
            "get": {
                "tags": [
                    "Aula"
                ],
                "summary": "Obtener todas las aulas",
                "description": "Devuelve todas las aulas",
                "operationId": "getAulas",
                "responses": {
                    "200": {
                        "description": "Aulas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Aula"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener las aulas"
                    }
                }
            }
        },
        "/api/aulas/laboratorios": {
            "get": {
                "tags": [
                    "Aula"
                ],
                "summary": "Obtener todos los laboratorios",
                "description": "Devuelve todos los laboratorios",
                "operationId": "getLaboratorios",
                "responses": {
                    "200": {
                        "description": "Laboratorios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Aula"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener los laboratorios"
                    }
                }
            }
        },
        "/api/aulas/{nro}": {
            "get": {
                "tags": [
                    "Aula"
                ],
                "summary": "Obtener aula por nro",
                "description": "Devuelve un aula",
                "operationId": "getAulaPorNro",
                "parameters": [
                    {
                        "name": "nro",
                        "in": "path",
                        "description": "Número del aula",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aula",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Aula"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No existe el aula"
                    },
                    "500": {
                        "description": "Error al obtener el aula"
                    }
                }
            }
        },
        "/api/aulas/guardar": {
            "post": {
                "tags": [
                    "Aula"
                ],
                "summary": "Guardar aula",
                "description": "Guardar un aula",
                "operationId": "guardarAula",
                "requestBody": {
                    "description": "Aula a guardar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AulaData"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Aula",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Aula"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al guardar el aula"
                    }
                }
            }
        },
        "/api/aulas/actualizar/{nro}": {
            "put": {
                "tags": [
                    "Aula"
                ],
                "summary": "Actualizar aula",
                "description": "Actualizar un aula",
                "operationId": "actualizarAula",
                "parameters": [
                    {
                        "name": "nro",
                        "in": "path",
                        "description": "Número del aula",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Aula a actualizar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AulaData"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Aula",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Aula"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No existe el aula"
                    },
                    "500": {
                        "description": "Error al actualizar el aula"
                    }
                }
            }
        },
        "/api/aulas/eliminar/{nro}": {
            "delete": {
                "tags": [
                    "Aula"
                ],
                "summary": "Eliminar aula por nro",
                "description": "Eliminar aula por nro",
                "operationId": "eliminarAulaPorNro",
                "parameters": [
                    {
                        "name": "nro",
                        "in": "path",
                        "description": "Número del aula",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aula eliminada correctamente"
                    },
                    "404": {
                        "description": "No existe el aula"
                    },
                    "500": {
                        "description": "Error al eliminar el aula"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AulaData": {
                "title": "AulaData",
                "description": "Esquema del objeto AulaData",
                "properties": {
                    "nro": {
                        "description": "Número del aula",
                        "type": "integer"
                    },
                    "laboratorio": {
                        "description": "Laboratorio del aula",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "Aula": {
                "title": "Aula",
                "description": "Esquema del objeto Aula",
                "properties": {
                    "nro": {
                        "description": "Número del aula",
                        "type": "integer"
                    },
                    "laboratorio": {
                        "description": "Laboratorio del aula",
                        "type": "boolean"
                    },
                    "fecha_creacion": {
                        "description": "Fecha de creacion del aula",
                        "type": "string",
                        "format": "date-time"
                    },
                    "fecha_modificacion": {
                        "description": "Fecha de modificacion del aula",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    }
}